requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.12.0
httpx==0.25.2
chromadb==0.4.15
sentence-transformers==2.7.0
huggingface-hub==0.20.3
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
numpy==1.24.3
